Don't start new ssh-agent if X session starts one
Either Ubuntu or Xfce starts an instance of ssh-agent on login; this
commit alters the ssh-agent module to not start a new process if one
already exists. If an ssh-agent process exists and the environment
variables are available then we create the prezto/ssh-agent session file
by hand and load identities normally.